chore(node): move API related files to `api` folder
