+++ packages/rolldown/src/api/build.ts
@@ -1,6 +1,6 @@
 import type { RolldownOptions } from '../types/rolldown-options'
 import type { RolldownOutput } from '../types/rolldown-output'
-import { rolldown } from '../rolldown'
+import { rolldown } from './rolldown'
 
 export async function build(options: RolldownOptions): Promise<RolldownOutput> {
   const { output, ...inputOptions } = options
