+++ packages/rolldown/src/experimental-index.ts
@@ -1,5 +1,5 @@
 export { defineParallelPlugin } from './plugin/parallel-plugin'
-export { experimental_scan as scan } from './rolldown'
+export { experimental_scan as scan } from './api/experimental'
 export { transform } from './binding'
 export type { TransformOptions, TransformResult } from './binding'
 export { composeJsPlugins as composePlugins } from './utils/compose-js-plugins'
